+++ pygsp/filters/__init__.py
@@ -55,6 +55,13 @@ Then, derived classes implement various common graph filters.
 
     Heat
     Expwin
+    Rectangular
+
+**Band pass filters**
+
+.. autosummary::
+
+    Rectangular
 
 Approximations
 --------------
@@ -97,6 +104,7 @@ _FILTERS = [
     'MexicanHat',
     'Meyer',
     'Papadakis',
+    'Rectangular',
     'Regular',
     'Simoncelli',
     'SimpleTight',