+++ doc/history.rst
@@ -21,6 +21,7 @@ Additional features:
 
 * print(graph) and print(filters) now show valuable information.
 * Building a graph object is much faster.
+* New rectangular filter (low-pass and band-pass).
 
 0.5.1 (2017-12-15)
 ------------------